mdcode: isolate the action catalog encoding in kc_actions.ts
Actions are the one model construct with no Knowledge Catalog type. Every
other construct maps to a built-in system type under dataplex-types/global
(semantic-model / semantic-entity / semantic-metric entries, schema-join
links, schema / guidelines aspects); actions ride the model anchor's built-in
`overview` aspect as Markdown plus an embedded JSON block.

That encoding was spread across four files. Move it whole into a new leaf
module, kc_actions.ts, with a four-symbol surface:

  ACTIONS_OVERVIEW_MARKER     the JSON block delimiter
  ACTION_ANCHOR_ASPECT_TYPES  which aspect types the encoding attaches
  actionAnchorAspects()       IR -> the anchor's aspects
  readActions()               the anchor -> IR

knowledge_catalog.ts spreads actionAnchorAspects into the anchor's aspect
map, kc_converter.ts calls readActions, and deploy_knowledge_catalog.ts and
pull_kc.ts name the aspect types through ACTION_ANCHOR_ASPECT_TYPES instead
of the literal 'overview'. No call site names the encoding.

The module's header states what to change when a built-in action type ships:
an aspect-shaped type is a one-file edit, and an entry-per-action type
deletes the file and moves the emit/read beside the metric code. It imports
only ./ir and ../gcp/dataplex, and keeps a few local helper copies so it can
be swapped or deleted as a unit.

No behavior change: the emitted aspects, the pulled IR, and every warning are
identical.
